[WebGPU] Move SharedBufferPool into its own translation unit
Pure code motion, no behaviour change. SharedBufferPool sat inside
buffer_manager.h/.cc, which obscured that it is a peer of SessionBufferPool
rather than an implementation detail of the cache managers. The two are easy to
confuse and have opposite scopes, so the new header states the distinction:

  SessionBufferPool  per session,  whole generations, fed by Graph/GraphSimple
  SharedBufferPool   per context,  individual buffers, fed by Bucket/Simple

The two serve disjoint cache managers, so nothing is shared between them.

NormalizeBufferSize and BUCKET_DEFAULT_LIMIT_TABLE move to the new header as
well: both encode buffer-sizing policy that the pool owns, and both are still
needed by the cache managers in buffer_manager.cc.

Note for anyone adding files here: onnxruntime_providers_webgpu.cmake picks up
sources with file(GLOB_RECURSE ... CONFIGURE_DEPENDS), which the Visual Studio
generator does not reliably re-run. A plain incremental build silently skipped
the new .cc and still reported success, because a static library does not
resolve symbols; the failure would only have appeared when linking an
executable. Re-run the configure step (build.py --update) after adding sources.

Verified: onnxruntime_provider_test links with no undefined symbols, and the 5
concurrency cases pass 3/3 each.
